refactor tag-release-track
tag-release-track script was broken. It was checking only if the
releast track tag was linked to a digest on the remote repository
but not checking if that digest was the correct one, the one for
the latest image.

With the refactor the script checks that local and remote digests
linked to a version and a release track are as expected.

The version to use as release track is based on the filesystem
sorting, as is using a simple algorithm that collects all folders
whose names start with the release track portion.